Updated README to reflect correct delete function
I believe the correct delete function is `model.deleteFromContext()` rather than `model.delete()` where model is an instantiated NSManagedObject subclass. I looked through the source to find a reference to `delete()` and `deleteFromContext()` was the closest I came to finding something similar. 

If you wanted to implement the `delete()` function for simplicities sake, you could just map it directly to `deleteFromContext()` given that the `context` var is optional and automatically assigned to the default context if not set.